Implement BASIC SDL2 support, change setup
Use two perspectives to repesent and draw the grid on a window,
currently only working for my monitor resolution, but will be fixed to
change that. We render all live cells in white. Also change
functionality of setup such that it no longer prints initial grid in num
representation and grid representation